requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
streamlit
python-dotenv
sqlalchemy
mysql-connector-python